fixed client tests
