+++ app/react/Viewer/actions/documentActions.js
@@ -57,7 +57,7 @@ export function saveDocument(doc) {
 export function saveToc(toc) {
   return function (dispatch, getState) {
     const {_id, _rev, sharedId} = getState().documentViewer.doc.toJS();
-    dispatch(formActions.reset('documentViewer.tocForm'));
+    dispatch(formActions.reset('documentViewer.sidepanel.metadata'));
     dispatch(actions.set('documentViewer/tocBeingEdited', false));
     return dispatch(saveDocument({_id, _rev, sharedId, toc}));
   };
